advection_diffusion_gpu.py
"""
GPU-Accelerated Advection-Diffusion Solver
===========================================

CuPy-based implementation for continuum colloidal concentration.
"""

try:
    import cupy as cp
    GPU_AVAILABLE = True
except ImportError:
    import numpy as cp
    GPU_AVAILABLE = False

import numpy as np
from typing import Tuple


class AdvectionDiffusionSolverGPU:
    """
    GPU-accelerated advection-diffusion solver.
    
    Speedup on A100: ~20-50x for large grids.
    """
    
    def __init__(
        self,
        nx: int,
        ny: int,
        size_x: float,
        size_y: float,
        dt: float,
        max_steps: int,
        tol: float
    ):
        self.nx = nx
        self.ny = ny
        self.size_x = size_x
        self.size_y = size_y
        self.dt = dt
        self.max_steps = max_steps
        self.tol = tol
        
        self.dx = size_x / nx
        self.dy = size_y / ny
        
        print(f"GPU AD Solver initialized: {nx}x{ny} grid")
    
    def solve(
        self,
        Dx_C1: np.ndarray,
        Dx_C2: np.ndarray,
        Dy_C1: np.ndarray,
        Dy_C2: np.ndarray,
        Pe: float,
        eta: float,
        alpha2: float,
        A_rd: float,
        Mob_C1: float,
        Mob_C2: float,
        Mob2_C1: float,
        Mob2_C2: float,
        N: int,
        N1: int,
        area_fraction: float
    ) -> Tuple[np.ndarray, np.ndarray]:
        """
        Solve on GPU, return results on CPU.
        """
        # Transfer input arrays to GPU
        if GPU_AVAILABLE:
            Dx_C1_gpu = cp.asarray(Dx_C1)
            Dx_C2_gpu = cp.asarray(Dx_C2)
            Dy_C1_gpu = cp.asarray(Dy_C1)
            Dy_C2_gpu = cp.asarray(Dy_C2)
        else:
            Dx_C1_gpu = Dx_C1
            Dx_C2_gpu = Dx_C2
            Dy_C1_gpu = Dy_C1
            Dy_C2_gpu = Dy_C2
        
        # Effective diffusion
        DNi = alpha2 * (Mob_C1 - Mob_C2 * eta * (1 + A_rd * eta) / A_rd) / Pe
        DN1 = DNi
        DN2 = DNi
        
        print(f"DNi = {DNi:.6e}")
        
        # Compute velocity fields on GPU
        Vx_dp1 = Mob_C1 * Dx_C1_gpu + Mob_C2 * Dx_C2_gpu
        Vy_dp1 = Mob_C1 * Dy_C1_gpu + Mob_C2 * Dy_C2_gpu
        Vx_dp2 = Mob2_C1 * Dx_C1_gpu + Mob2_C2 * Dx_C2_gpu
        Vy_dp2 = Mob2_C1 * Dy_C1_gpu + Mob2_C2 * Dy_C2_gpu
        
        # Initialize volume fractions on GPU
        phi1_n = cp.ones((self.nx + 2, self.ny + 2)) * (N1 / N) * area_fraction
        phi2_n = cp.ones((self.nx + 2, self.ny + 2)) * ((N - N1) / N) * area_fraction
        
        phi1_n_1 = phi1_n.copy()
        phi2_n_1 = phi2_n.copy()
        
        # Time stepping on GPU
        for step in range(1, self.max_steps + 1):
            if step == 1:
                F1_n = self._compute_rhs_gpu(phi1_n, Vx_dp1, Vy_dp1, DN1)
                F2_n = self._compute_rhs_gpu(phi2_n, Vx_dp2, Vy_dp2, DN2)
                
                phi1_new = phi1_n + self.dt * F1_n
                phi2_new = phi2_n + self.dt * F2_n
            else:
                F1_n_1 = self._compute_rhs_gpu(phi1_n_1, Vx_dp1, Vy_dp1, DN1)
                F1_n = self._compute_rhs_gpu(phi1_n, Vx_dp1, Vy_dp1, DN1)
                F2_n_1 = self._compute_rhs_gpu(phi2_n_1, Vx_dp2, Vy_dp2, DN2)
                F2_n = self._compute_rhs_gpu(phi2_n, Vx_dp2, Vy_dp2, DN2)
                
                phi1_new = phi1_n + self.dt * (1.5 * F1_n - 0.5 * F1_n_1)
                phi2_new = phi2_n + self.dt * (1.5 * F2_n - 0.5 * F2_n_1)
                
                phi1_new = self._apply_periodic_bc_gpu(phi1_new)
                phi2_new = self._apply_periodic_bc_gpu(phi2_new)
                
                errorX = float(cp.mean(cp.abs(phi1_new - phi1_n)))
                errorY = float(cp.mean(cp.abs(phi2_new - phi2_n)))
                
                if step % 10000 == 0:
                    print(f"AD Step {step}: error = {errorX:.6e}, {errorY:.6e}")
                
                if errorX < self.tol and errorY < self.tol:
                    print(f"Converged at step {step}")
                    break
            
            phi1_n_1 = phi1_n.copy()
            phi2_n_1 = phi2_n.copy()
            phi1_n = phi1_new
            phi2_n = phi2_new
        
        # Transfer results to CPU
        phi1 = cp.asnumpy(phi1_n[1:self.nx+2, 1:self.ny+2]) if GPU_AVAILABLE else phi1_n[1:self.nx+2, 1:self.ny+2]
        phi2 = cp.asnumpy(phi2_n[1:self.nx+2, 1:self.ny+2]) if GPU_AVAILABLE else phi2_n[1:self.nx+2, 1:self.ny+2]
        
        return phi1, phi2
    
    def _compute_rhs_gpu(
        self,
        phi: cp.ndarray,
        Vx: cp.ndarray,
        Vy: cp.ndarray,
        D: float
    ) -> cp.ndarray:
        """
        Compute RHS on GPU using vectorized operations.
        """
        F = cp.zeros_like(phi)
        
        # Vectorized upwind scheme for advection
        # Positive velocity: forward difference
        # Negative velocity: backward difference
        Vx_pos = cp.maximum(Vx, 0)
        Vx_neg = cp.minimum(Vx, 0)
        Vy_pos = cp.maximum(Vy, 0)
        Vy_neg = cp.minimum(Vy, 0)
        
        # Advection term (vectorized)
        dx_inv = 1.0 / self.dx
        dy_inv = 1.0 / self.dy
        
        advection_x = cp.zeros_like(phi)
        advection_y = cp.zeros_like(phi)
        
        # Interior points (vectorized slicing)
        advection_x[1:-1, 1:-1] = (
            Vx_pos[1:-1, 1:-1] * (phi[1:-1, 1:-1] - phi[:-2, 1:-1]) * dx_inv +
            Vx_neg[1:-1, 1:-1] * (phi[2:, 1:-1] - phi[1:-1, 1:-1]) * dx_inv
        )
        
        advection_y[1:-1, 1:-1] = (
            Vy_pos[1:-1, 1:-1] * (phi[1:-1, 1:-1] - phi[1:-1, :-2]) * dy_inv +
            Vy_neg[1:-1, 1:-1] * (phi[1:-1, 2:] - phi[1:-1, 1:-1]) * dy_inv
        )
        
        # Diffusion term (centered differences, vectorized)
        dx2_inv = 1.0 / (self.dx ** 2)
        dy2_inv = 1.0 / (self.dy ** 2)
        
        diffusion_x = cp.zeros_like(phi)
        diffusion_y = cp.zeros_like(phi)
        
        diffusion_x[1:-1, 1:-1] = D * (
            phi[2:, 1:-1] - 2*phi[1:-1, 1:-1] + phi[:-2, 1:-1]
        ) * dx2_inv
        
        diffusion_y[1:-1, 1:-1] = D * (
            phi[1:-1, 2:] - 2*phi[1:-1, 1:-1] + phi[1:-1, :-2]
        ) * dy2_inv
        
        F[1:-1, 1:-1] = -(advection_x[1:-1, 1:-1] + advection_y[1:-1, 1:-1]) + \
                        (diffusion_x[1:-1, 1:-1] + diffusion_y[1:-1, 1:-1])
        
        return F
    
    def _apply_periodic_bc_gpu(self, field: cp.ndarray) -> cp.ndarray:
        """Apply periodic BC on GPU (vectorized)."""
        nx, ny = self.nx, self.ny
        
        # Corners
        field[0, 0] = field[nx, ny]
        field[0, ny+1] = field[nx, 1]
        field[nx+1, 0] = field[1, ny]
        field[nx+1, ny+1] = field[1, 1]
        
        # Edges (vectorized operations)
        field[0, 1:ny+1] = field[nx, 1:ny+1]
        field[nx+1, 1:ny+1] = field[1, 1:ny+1]
        field[1:nx+1, 0] = field[1:nx+1, ny]
        field[1:nx+1, ny+1] = field[1:nx+1, 1]
        
        return field


if not GPU_AVAILABLE:
    print("Warning: CuPy not available. Import from advection_diffusion.py instead.")
    from advection_diffusion import AdvectionDiffusionSolver as AdvectionDiffusionSolverGPU
