requirements_gpu_cuda12.txt
# Requirements for GPU-Accelerated Simulation on CUDA 12
# Use this for Google Colab with A100 or systems with CUDA 12

# GPU Acceleration (CUDA 12)
cupy-cuda12x>=12.0.0

# Core numerical computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0

# Performance optimization
numba>=0.54.0

# Animation support (optional)
ffmpeg-python>=0.2.0

# Development tools (optional)
pytest>=6.2.0
pytest-cov>=2.12.0

# Jupyter support for Colab
jupyter>=1.0.0
ipywidgets>=7.6.0

# ============================================
# Alternative CuPy installations:
# ============================================
# For CUDA 11.x:
# cupy-cuda11x>=11.0.0

# For CPU-only (fallback):
# Comment out cupy lines above

# ============================================
# Installation order for Colab:
# ============================================
# 1. !pip install cupy-cuda12x
# 2. !pip install scipy matplotlib numba
# 3. (Optional) !pip install ffmpeg-python

# ============================================
# Verify GPU installation:
# ============================================
# import cupy as cp
# print(cp.cuda.Device())
# print("GPU Memory:", cp.cuda.Device().mem_info)
